requirements.txt
requests

configparser

kafka-python